+++ src/test/converter2/issues/gh3064.ts
@@ -1,0 +1,11 @@
+/**
+ * A
+ */
+export type ErrorResult = 1 | 2;
+
+/**
+ * B
+ */
+export function ErrorResult(err: any): 1 {
+    return 1;
+}