+++ CHANGELOG.md
@@ -6,6 +6,7 @@ title: Changelog
 
 ### Bug Fixes
 
+- Improved handling of comments for type aliases which have been declaration merged with functions, #3064.
 - Fixed anchor link generation to members named `$`, #3065.
 
 ## v0.28.16 (2026-01-12)